backend/requirements.txt
flask
flask-cors
requests
pydub
openai-whisper
torch
torchaudio
numpy
scipy
librosa
soundfile
